+++ client/browser/src/libs/code_intelligence/code_intelligence.tsx
@@ -16,12 +16,14 @@ import { toPrettyBlobURL } from '@sourcegraph/codeintellify/lib/url'
 import * as H from 'history'
 import * as React from 'react'
 import { createPortal, render } from 'react-dom'
-import { animationFrameScheduler, BehaviorSubject, Observable, of, Subject, Subscription, Unsubscribable } from 'rxjs'
+import { animationFrameScheduler, Observable, of, Subject, Subscription, Unsubscribable } from 'rxjs'
 import { filter, map, mergeMap, observeOn, withLatestFrom } from 'rxjs/operators'
 
-import { Environment } from '../../../../../shared/src/api/client/environment'
+import { Model } from '../../../../../shared/src/api/client/model'
 import { TextDocumentItem } from '../../../../../shared/src/api/client/types/textDocument'
+import { ExtensionsControllerProps } from '../../../../../shared/src/extensions/controller'
 import { getModeFromPath } from '../../../../../shared/src/languages'
+import { PlatformContextProps } from '../../../../../shared/src/platform/context'
 import {
     createJumpURLFetcher,
     createLSPFromExtensions,
@@ -37,7 +39,7 @@ import { githubCodeHost } from '../github/code_intelligence'
 import { gitlabCodeHost } from '../gitlab/code_intelligence'
 import { phabricatorCodeHost } from '../phabricator/code_intelligence'
 import { findCodeViews, getContentOfCodeView } from './code_views'
-import { applyDecoration, Controllers, initializeExtensions } from './extensions'
+import { applyDecoration, initializeExtensions } from './extensions'
 
 /**
  * Defines a type of code view a given code host can have. It tells us how to
@@ -207,16 +209,13 @@ export interface FileInfo {
  * @param codeHost
  */
 function initCodeIntelligence(
-    codeHost: CodeHost,
-    environment: BehaviorSubject<Pick<Environment, 'roots' | 'visibleTextDocuments'>>
+    codeHost: CodeHost
 ): {
     hoverifier: Hoverifier
-    controllers: Partial<Controllers>
+    controllers: Partial<ExtensionsControllerProps & PlatformContextProps>
 } {
-    const { platformContext, extensionsController }: Partial<Controllers> =
-        useExtensions && codeHost.getCommandPaletteMount
-            ? initializeExtensions(codeHost.getCommandPaletteMount, environment)
-            : {}
+    const { platformContext, extensionsController }: Partial<ExtensionsControllerProps & PlatformContextProps> =
+        useExtensions && codeHost.getCommandPaletteMount ? initializeExtensions(codeHost.getCommandPaletteMount) : {}
     const simpleProviderFns = extensionsController ? createLSPFromExtensions(extensionsController) : lspViaAPIXlang
 
     /** Emits when the go to definition button was clicked */
@@ -387,14 +386,10 @@ export interface ResolvedCodeView extends CodeViewWithOutSelector {
 }
 
 function handleCodeHost(codeHost: CodeHost): Subscription {
-    const environmentSubject = new BehaviorSubject<Pick<Environment, 'roots' | 'visibleTextDocuments'>>({
-        roots: null,
-        visibleTextDocuments: null,
-    })
     const {
         hoverifier,
         controllers: { platformContext, extensionsController },
-    } = initCodeIntelligence(codeHost, environmentSubject)
+    } = initCodeIntelligence(codeHost)
 
     const subscriptions = new Subscription()
 
@@ -465,7 +460,7 @@ function handleCodeHost(codeHost: CodeHost): Subscription {
                                 text: content!,
                             },
                         ]
-                        const roots: Environment['roots'] = [{ uri: toRootURI(info) }]
+                        const roots: Model['roots'] = [{ uri: toRootURI(info) }]
 
                         // When codeView is a diff, add BASE too.
                         if (baseContent! && info.baseRepoPath && info.baseCommitID && info.baseFilePath) {
@@ -511,7 +506,7 @@ function handleCodeHost(codeHost: CodeHost): Subscription {
                                 })
                         }
 
-                        environmentSubject.next({ roots, visibleTextDocuments })
+                        extensionsController.services.model.model.next({ roots, visibleTextDocuments })
                     }
 
                     const resolveContext: ContextResolver = ({ part }) => ({