+++ shared/src/api/integration-test/context.test.ts
@@ -1,19 +1,14 @@
 import * as assert from 'assert'
-import { distinctUntilChanged, map } from 'rxjs/operators'
+import { from } from 'rxjs'
+import { distinctUntilChanged } from 'rxjs/operators'
 import { ContextValues } from 'sourcegraph'
 import { collectSubscribableValues, integrationTestContext } from './helpers.test'
 
 describe('Context (integration)', () => {
     describe('internal.updateContext', () => {
         it('updates context', async () => {
-            const { clientController, extensionHost, ready } = await integrationTestContext()
-            await ready
-            const values = collectSubscribableValues(
-                clientController.environment.pipe(
-                    map(({ context }) => context),
-                    distinctUntilChanged()
-                )
-            )
+            const { services, extensionHost } = await integrationTestContext()
+            const values = collectSubscribableValues(from(services.context.data).pipe(distinctUntilChanged()))
 
             extensionHost.internal.updateContext({ a: 1 })
             await extensionHost.internal.sync()