+++ shared/src/api/integration-test/windows.test.ts
@@ -1,15 +1,14 @@
 import assert from 'assert'
 import { map } from 'rxjs/operators'
 import { Window } from 'sourcegraph'
+import { MessageType } from '../client/services/notifications'
 import { assertToJSON } from '../extension/types/common.test'
-import { MessageType } from '../protocol'
 import { collectSubscribableValues, integrationTestContext } from './helpers.test'
 
 describe('Windows (integration)', () => {
     describe('app.activeWindow', () => {
         it('returns the active window', async () => {
-            const { extensionHost, ready } = await integrationTestContext()
-            await ready
+            const { extensionHost } = await integrationTestContext()
             assertToJSON(extensionHost.app.activeWindow, {
                 visibleViewComponents: [
                     {
@@ -23,8 +22,7 @@ describe('Windows (integration)', () => {
 
     describe('app.windows', () => {
         it('lists windows', async () => {
-            const { extensionHost, ready } = await integrationTestContext()
-            await ready
+            const { extensionHost } = await integrationTestContext()
             assertToJSON(extensionHost.app.windows, [
                 {
                     visibleViewComponents: [
@@ -38,15 +36,14 @@ describe('Windows (integration)', () => {
         })
 
         it('adds new text documents', async () => {
-            const { clientController, extensionHost, getEnvironment, ready } = await integrationTestContext()
+            const { model, extensionHost } = await integrationTestContext()
 
-            const prevEnvironment = getEnvironment()
-            clientController.setEnvironment({
-                ...prevEnvironment,
+            model.next({
+                ...model.value,
                 visibleTextDocuments: [{ uri: 'file:///f2', languageId: 'l2', text: 't2' }],
             })
+            await extensionHost.internal.sync()
 
-            await ready
             assertToJSON(extensionHost.app.windows, [
                 {
                     visibleViewComponents: [
@@ -62,31 +59,28 @@ describe('Windows (integration)', () => {
 
     describe('Window', () => {
         it('Window#showNotification', async () => {
-            const { clientController, extensionHost, ready } = await integrationTestContext()
-            await ready
-            const values = collectSubscribableValues(clientController.showMessages)
+            const { extensionHost, services } = await integrationTestContext()
+            const values = collectSubscribableValues(services.notifications.showMessages)
             extensionHost.app.activeWindow!.showNotification('a') // tslint:disable-line deprecation
             await extensionHost.internal.sync()
             assert.deepStrictEqual(values, [{ message: 'a', type: MessageType.Info }] as typeof values)
         })
 
         it('Window#showMessage', async () => {
-            const { clientController, extensionHost, ready } = await integrationTestContext()
-            clientController.showMessageRequests.subscribe(({ resolve }) => resolve(Promise.resolve(null)))
-            await ready
+            const { extensionHost, services } = await integrationTestContext()
+            services.notifications.showMessageRequests.subscribe(({ resolve }) => resolve(Promise.resolve(null)))
             const values = collectSubscribableValues(
-                clientController.showMessageRequests.pipe(map(({ message, type }) => ({ message, type })))
+                services.notifications.showMessageRequests.pipe(map(({ message, type }) => ({ message, type })))
             )
             assert.strictEqual(await extensionHost.app.activeWindow!.showMessage('a'), null)
             assert.deepStrictEqual(values, [{ message: 'a', type: MessageType.Info }] as typeof values)
         })
 
         it('Window#showInputBox', async () => {
-            const { clientController, extensionHost, ready } = await integrationTestContext()
-            clientController.showInputs.subscribe(({ resolve }) => resolve(Promise.resolve('c')))
-            await ready
+            const { extensionHost, services } = await integrationTestContext()
+            services.notifications.showInputs.subscribe(({ resolve }) => resolve(Promise.resolve('c')))
             const values = collectSubscribableValues(
-                clientController.showInputs.pipe(map(({ message, defaultValue }) => ({ message, defaultValue })))
+                services.notifications.showInputs.pipe(map(({ message, defaultValue }) => ({ message, defaultValue })))
             )
             assert.strictEqual(await extensionHost.app.activeWindow!.showInputBox({ prompt: 'a', value: 'b' }), 'c')
             assert.deepStrictEqual(values, [{ message: 'a', defaultValue: 'b' }] as typeof values)