+++ shared/src/extensions/controller.ts
@@ -1,30 +1,31 @@
-import { from, Subject, Unsubscribable } from 'rxjs'
-import { filter, first, map, mergeMap } from 'rxjs/operators'
-import { Controller as BaseController, ExtensionConnectionKey } from '../api/client/controller'
-import { Environment } from '../api/client/environment'
+import { combineLatest, from, Observable, Subject, Subscription, Unsubscribable } from 'rxjs'
+import { distinctUntilChanged, map, share, switchMap, tap } from 'rxjs/operators'
+import { createExtensionHostClient } from '../api/client/client'
+import { Services } from '../api/client/services'
 import { ExecuteCommandParams } from '../api/client/services/command'
-import { Contributions, MessageType } from '../api/protocol'
+import { ContributionRegistry } from '../api/client/services/contribution'
+import { ExtensionsService } from '../api/client/services/extensionsService'
+import { MessageType } from '../api/client/services/notifications'
+import { InitData } from '../api/extension/extensionHost'
+import { Contributions } from '../api/protocol'
+import { createConnection } from '../api/protocol/jsonrpc2/connection'
 import { BrowserConsoleTracer } from '../api/protocol/jsonrpc2/trace'
-import { registerBuiltinClientCommands, updateConfiguration } from '../commands/commands'
+import { registerBuiltinClientCommands } from '../commands/commands'
 import { Notification } from '../notifications/notification'
 import { PlatformContext } from '../platform/context'
-import { ExtensionManifest } from '../schema/extension.schema'
-import { SettingsCascade } from '../settings/settings'
-import { asError, isErrorLike } from '../util/errors'
-import { ConfiguredExtension, isExtensionEnabled } from './extension'
-import { ExtensionStatus } from './ExtensionStatus'
+import { isErrorLike } from '../util/errors'
+import { ExtensionManifest } from './extensionManifest'
 
-/**
- * Extends the {@link BaseController} class to add functionality that is useful to this package's consumers.
- */
-export class Controller extends BaseController<ConfiguredExtension, SettingsCascade> {
+export interface Controller extends Unsubscribable {
     /**
      * Global notification messages that should be displayed to the user, from the following sources:
      *
      * - window/showMessage notifications from extensions
      * - Errors thrown or returned in command invocation
      */
-    public readonly notifications = new Subject<Notification>()
+    readonly notifications: Observable<Notification>
+
+    services: Services
 
     /**
      * Executes the command (registered in the CommandRegistry) specified in params. If an error is thrown, the
@@ -34,181 +35,133 @@ export class Controller extends BaseController<ConfiguredExtension, SettingsCasc
      * {@link sourcegraph:CommandRegistry#executeCommand} directly (to ensure errors are
      * emitted as notifications).
      */
-    public executeCommand(params: ExecuteCommandParams): Promise<any> {
-        return this.services.commands.executeCommand(params).catch(err => {
-            this.notifications.next({ message: err, type: MessageType.Error, source: params.command })
-            return Promise.reject(err)
-        })
-    }
+    executeCommand(params: ExecuteCommandParams): Promise<any>
+
+    /**
+     * Frees all resources associated with this client.
+     */
+    unsubscribe(): void
 }
 
 /**
- * React props or state containing the controller. There should be only a single controller for the whole
+ * React props or state containing the client. There should be only a single client for the whole
  * application.
  */
 export interface ExtensionsControllerProps {
     /**
-     * The controller, which is used to communicate with the extensions and manages extensions based on the
-     * environment.
+     * The client, which is used to communicate with and manage extensions.
      */
     extensionsController: Controller
 }
 
 /**
- * Filter the environment to omit extensions that should not be activated (based on their manifest's
- * activationEvents).
- */
-function environmentFilter(
-    nextEnvironment: Environment<ConfiguredExtension, SettingsCascade>
-): Environment<ConfiguredExtension, SettingsCascade> {
-    return {
-        ...nextEnvironment,
-        extensions:
-            nextEnvironment.extensions &&
-            nextEnvironment.extensions.filter(x => {
-                try {
-                    if (!isExtensionEnabled(nextEnvironment.configuration.final, x.id)) {
-                        return false
-                    } else if (!x.manifest) {
-                        console.warn(
-                            `Extension ${x.id} was not found. Remove it from settings to suppress this warning.`
-                        )
-                        return false
-                    } else if (isErrorLike(x.manifest)) {
-                        console.warn(asError(x.manifest))
-                        return false
-                    } else if (!x.manifest.activationEvents) {
-                        console.warn(`Extension ${x.id} has no activation events, so it will never be activated.`)
-                        return false
-                    }
-                    const visibleTextDocumentLanguages = nextEnvironment.visibleTextDocuments
-                        ? nextEnvironment.visibleTextDocuments.map(({ languageId }) => languageId)
-                        : []
-                    return x.manifest.activationEvents.some(
-                        e => e === '*' || visibleTextDocumentLanguages.some(l => e === `onLanguage:${l}`)
-                    )
-                } catch (err) {
-                    console.error(err)
-                }
-                return false
-            }),
-    }
-}
-
-declare global {
-    interface Window {
-        sx: any
-        sxenv: any
-    }
-}
-
-/**
- * Creates the controller, which handles all communication between the React app and Sourcegraph extensions.
+ * Creates the controller, which handles all communication between the client application and extensions.
  *
- * There should only be a single controller for the entire application. The controller's environment represents all
- * of the application state that the controller needs to know.
- *
- * It receives state updates via calls to the setEnvironment method from React components. It provides results to
- * React components via its services and the showMessages, etc., observables.
+ * There should only be a single controller for the entire client application. The controller's model represents
+ * all of the client application state that the client needs to know.
  */
 export function createController(context: PlatformContext): Controller {
-    const controller: Controller = new Controller({
-        clientOptions: (_key: ExtensionConnectionKey, extension: ConfiguredExtension) => ({
-            createMessageTransports: async () => {
-                const settingsCascade = await controller.environment
-                    .pipe(
-                        first(),
-                        map(({ configuration }) => configuration)
-                    )
-                    .toPromise()
-                return context.createMessageTransports(extension, settingsCascade)
-            },
-        }),
-        environmentFilter,
-    })
+    const subscriptions = new Subscription()
 
-    // Apply trace settings.
-    //
-    // HACK(sqs): This is inefficient and doesn't unsubscribe itself.
-    controller.clientEntries.subscribe(entries => {
-        const traceEnabled = localStorage.getItem(ExtensionStatus.TRACE_STORAGE_KEY) !== null
-        for (const e of entries) {
-            e.connection
-                .then(c => c.trace(traceEnabled ? new BrowserConsoleTracer(e.key.id) : null))
-                .catch(err => console.error(err))
-        }
-    })
+    const services = new Services(context)
+    const extensionHostConnection = combineLatest(
+        context.createExtensionHost().pipe(
+            switchMap(async messageTransports => {
+                const connection = createConnection(messageTransports)
+                connection.listen()
 
-    registerBuiltinClientCommands(context, controller)
-    registerExtensionContributions(controller)
+                const initData: InitData = {
+                    sourcegraphURL: context.sourcegraphURL,
+                    clientApplication: context.clientApplication,
+                }
+                await connection.sendRequest('initialize', [initData])
+                return connection
+            }),
+            share()
+        ),
+        context.traceExtensionHostCommunication
+    ).pipe(
+        tap(([connection, trace]) => connection.trace(trace ? new BrowserConsoleTracer('') : null)),
+        map(([connection]) => connection),
+        distinctUntilChanged()
+    )
+    const client = createExtensionHostClient(services, extensionHostConnection)
+    subscriptions.add(client)
+
+    const notifications = new Subject<Notification>()
+
+    subscriptions.add(registerBuiltinClientCommands(services, context))
+    subscriptions.add(registerExtensionContributions(services.contribution, services.extensions))
 
     // Show messages (that don't need user input) as global notifications.
-    controller.showMessages.subscribe(({ message, type }) => controller.notifications.next({ message, type }))
+    subscriptions.add(
+        services.notifications.showMessages.subscribe(({ message, type }) => notifications.next({ message, type }))
+    )
 
     function messageFromExtension(message: string): string {
         return `From extension:\n\n${message}`
     }
-    controller.showMessageRequests.subscribe(({ message, actions, resolve }) => {
-        if (!actions || actions.length === 0) {
-            alert(messageFromExtension(message))
-            resolve(null)
-            return
-        }
-        const value = prompt(
-            messageFromExtension(
-                `${message}\n\nValid responses: ${actions.map(({ title }) => JSON.stringify(title)).join(', ')}`
-            ),
-            actions[0].title
-        )
-        resolve(actions.find(a => a.title === value) || null)
-    })
-    controller.showInputs.subscribe(({ message, defaultValue, resolve }) =>
-        resolve(prompt(messageFromExtension(message), defaultValue))
+    subscriptions.add(
+        services.notifications.showMessageRequests.subscribe(({ message, actions, resolve }) => {
+            if (!actions || actions.length === 0) {
+                alert(messageFromExtension(message))
+                resolve(null)
+                return
+            }
+            const value = prompt(
+                messageFromExtension(
+                    `${message}\n\nValid responses: ${actions.map(({ title }) => JSON.stringify(title)).join(', ')}`
+                ),
+                actions[0].title
+            )
+            resolve(actions.find(a => a.title === value) || null)
+        })
     )
-    controller.configurationUpdates
-        .pipe(
-            mergeMap(params => {
-                const update = updateConfiguration(context, params)
-                params.resolve(update)
-                return from(update)
-            })
+    subscriptions.add(
+        services.notifications.showInputs.subscribe(({ message, defaultValue, resolve }) =>
+            resolve(prompt(messageFromExtension(message), defaultValue))
         )
-        .subscribe(undefined, err => console.error(err))
+    )
 
     // Print window/logMessage log messages to the browser devtools console.
-    controller.logMessages.subscribe(({ message }) => {
-        log('info', 'EXT', message)
-    })
+    subscriptions.add(
+        services.notifications.logMessages.subscribe(({ message }) => {
+            log('info', 'EXT', message)
+        })
+    )
 
     // Debug helpers.
     const DEBUG = true
     if (DEBUG) {
-        // Debug helper: log environment changes.
-        const LOG_ENVIRONMENT = false
-        if (LOG_ENVIRONMENT) {
-            controller.environment.subscribe(environment => log('info', 'env', environment))
+        // Debug helper: log model changes.
+        const LOG_MODEL = false
+        if (LOG_MODEL) {
+            subscriptions.add(services.model.model.subscribe(model => log('info', 'model', model)))
         }
 
-        // Debug helpers: e.g., just run `sx` in devtools to get a reference to this controller. (If multiple
-        // controllers are created, this points to the last one created.)
-        window.sx = controller
-        // This value is synchronously available because observable has an underlying
-        // BehaviorSubject source.
-        controller.environment.subscribe(v => (window.sxenv = v))
+        // Debug helpers: e.g., just run `sxservices` in devtools to get a reference to the services.
+        ;(window as any).sxservices = services
+        // This value is synchronously available because observable has an underlying BehaviorSubject source.
+        subscriptions.add(services.model.model.subscribe(v => ((window as any).sxmodel = v)))
     }
 
-    return controller
+    return {
+        notifications,
+        services,
+        executeCommand: params =>
+            services.commands.executeCommand(params).catch(err => {
+                notifications.next({ message: err, type: MessageType.Error, source: params.command })
+                return Promise.reject(err)
+            }),
+        unsubscribe: () => subscriptions.unsubscribe(),
+    }
 }
 
-/**
- * Registers the builtin client commands that are required by Sourcegraph extensions. See
- * {@link module:sourcegraph.module/protocol/contribution.ActionContribution#command} for
- * documentation.
- */
-function registerExtensionContributions(controller: Controller): Unsubscribable {
-    const contributions = controller.environment.pipe(
-        map(({ extensions }) => extensions),
-        filter((extensions): extensions is ConfiguredExtension[] => !!extensions),
+function registerExtensionContributions(
+    contributionRegistry: ContributionRegistry,
+    { enabledExtensions }: Pick<ExtensionsService, 'enabledExtensions'>
+): Unsubscribable {
+    const contributions = from(enabledExtensions).pipe(
         map(extensions =>
             extensions
                 .map(({ manifest }) => manifest)
@@ -217,7 +170,7 @@ function registerExtensionContributions(controller: Controller): Unsubscribable 
                 .filter((contributions): contributions is Contributions => !!contributions)
         )
     )
-    return controller.services.contribution.registerContributions({
+    return contributionRegistry.registerContributions({
         contributions,
     })
 }